update: move sort to types catalog
