+++ server/src/types/sortType.ts
@@ -1,0 +1,10 @@
+enum SortType {
+  NEWEST = 'newest',
+  OLDEST = 'oldest',
+  ALPHABETIC = 'alphabetic',
+  ALPHABETIC_REVERSE = 'alphabeticReverse',
+  LOWEST_PRICE = 'lowestPrice',
+  HIGHEST_PRICE = 'highestPrice'
+}
+
+export default SortType;
